Extract flow emission and SFC projection boundaries
