+++ src/test/scala/com/boombustgroup/amorfati/engine/flows/FlowSimulationSpec.scala
@@ -6,7 +6,7 @@ import com.boombustgroup.amorfati.tags.Heavy
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
-/** Tests FlowSimulation.emitAllBatches with MonthlyCalculus from
+/** Tests MonthFlowEmitter.emitAllBatches with MonthlyCalculus from
   * FlowSimulation.computeCalculus.
   *
   * Proves that the Contract-First pipeline design (MonthlyCalculus →
@@ -19,11 +19,11 @@ class FlowSimulationSpec extends AnyFlatSpec with Matchers:
 
   private given p: SimParams = SimParams.defaults
 
-  "FlowSimulation.emitAllBatches" should "match the step batch plan for computed calculus" in {
+  "MonthFlowEmitter.emitAllBatches" should "match the step batch plan for computed calculus" in {
     val state                   = stateFromSeed()
     val result                  = stepWithSeed(state)
     given RuntimeLedgerTopology = result.execution.topology
-    val flows                   = FlowSimulation.emitAllBatches(result.calculus)
+    val flows                   = MonthFlowEmitter.emitAllBatches(result.calculus)
 
     canonicalBatches(flows) shouldBe canonicalBatches(result.flows)
     result.execution.netDelta shouldBe 0L
